Added Rudementary Support for OpenCV
Added a simple C project to prototype the functionality of opencv apart
from the ARDrone API

Performed significant amount of work to clean up the ARDrone API to
make development more managable.

Added initial support for OpenCV in the ARDrone API.
- Added References to makefile to add in appropriate libraries
- Added Code in the Application which creates an opencv window and
connects to the camera on the computer itself

TODO -
Convert the video stream from the Drone into the OpenCV to be displayed
on the screen.